refactor(sdk)!: host-only audio-session ownership (remove dual mode)
The SDK no longer manages the AVAudioSession at all. Removes the
.sdkOwned/.hostOwned dual mode and all of its transitional machinery,
which was the source of repeated review churn (split-brain release
fallback, late-application ordering, double-configure divergence) and
had no value given a single consumer that migrates in lockstep.

Deleted: PlayolaAudioSessionOwnership, AudioSessionManager +
NoOpAudioSessionManager + AudioSessionManaging (whole file),
applyOwnership/appliedOwnership/sessionTouched/hasAppliedOwnershipToMixer,
handlesSessionEventsInternally, observer gating, and the legacy
interruption/route/engine-config handlers (removing the two known
pre-existing legacy bugs with them). configure() loses its
audioSessionOwnership parameter.

The host now owns the AVAudioSession: it configures/activates it and
drives pauseForInterruption()/resumeAfterInterruption() from its own
interruption handlers. SpinPlayer/PlayolaMainMixer no longer touch the
session; engine.start() throwing on an inactive session surfaces through
the normal error path.

Unchanged: pauseForInterruption/resumeAfterInterruption/.paused, lazy
mixer resolution, stale-download guard. The seam invariant test now
asserts the SDK references AVAudioSession nowhere in Sources/. Example
app gains its own session setup. README documents the host contract and
a 0.19->0.20 migration guide.

BREAKING CHANGE: host apps must own the AVAudioSession (see README
migration guide). Ships as 0.20.0.
