+++ Tests/PlayolaPlayerTests/InterruptionTransportTests.swift
@@ -1,0 +1,93 @@
+//  InterruptionTransportTests.swift
+//  PlayolaPlayer
+//
+//  Host-driven interruption transport: pauseForInterruption() /
+//  resumeAfterInterruption(). The SDK does not own the AVAudioSession or observe
+//  interruptions; the host calls these explicitly. None of these tests touch the
+//  shared CoreAudio graph: constructing a station player resolves the mixer
+//  lazily, and the resume tests exercise only the unarmed early-return path
+//  (the armed path would start a real engine).
+
+import Foundation
+import Testing
+
+@testable import PlayolaPlayer
+
+@MainActor
+struct InterruptionTransportTests {
+  @Test("pauseForInterruption bumps generation, publishes .paused, keeps station id")
+  func pauseBumpsGenerationPublishesPausedAndKeepsStationId() {
+    let player = PlayolaStationPlayer(
+      fileDownloadManager: MockFileDownloadManager(), urlSession: MockURLSession())
+    player.configure(authProvider: MockAuthProvider())
+    player.setStateForTesting(.playing(.mock), stationId: "station-1")
+    let generationBefore = player.playGeneration
+
+    player.pauseForInterruption()
+
+    #expect(player.playGeneration == generationBefore + 1)
+    #expect(player.stationId == "station-1")
+    #expect(player.isCurrentGeneration(generationBefore) == false)
+    if case .paused(let spin) = player.state {
+      #expect(spin.id == Spin.mock.id)
+    } else {
+      Issue.record("pause must publish .paused, got \(player.state)")
+    }
+  }
+
+  @Test("resumeAfterInterruption without prior pause is a no-op")
+  func resumeWithoutPriorPauseIsNoOp() async throws {
+    let player = PlayolaStationPlayer(
+      fileDownloadManager: MockFileDownloadManager(), urlSession: MockURLSession())
+    player.configure(authProvider: MockAuthProvider())
+    try await player.resumeAfterInterruption()
+    #expect(player.isPlaying == false)
+    #expect(player.interruptedStationIdForTesting == nil)
+  }
+
+  @Test("pause while not playing does not arm resume")
+  func pauseWhileNotPlayingDoesNotArmResume() async throws {
+    let player = PlayolaStationPlayer(
+      fileDownloadManager: MockFileDownloadManager(), urlSession: MockURLSession())
+    player.configure(authProvider: MockAuthProvider())
+    player.setStateForTesting(.idle, stationId: "station-1")
+
+    player.pauseForInterruption()
+    try await player.resumeAfterInterruption()
+
+    #expect(player.isPlaying == false)
+    #expect(player.interruptedStationIdForTesting == nil)
+  }
+
+  @Test("Double pause keeps resume armed")
+  func doublePauseKeepsResumeArmed() {
+    let player = PlayolaStationPlayer(
+      fileDownloadManager: MockFileDownloadManager(), urlSession: MockURLSession())
+    player.configure(authProvider: MockAuthProvider())
+    player.setStateForTesting(.playing(.mock), stationId: "station-1")
+
+    player.pauseForInterruption()
+    player.pauseForInterruption()
+
+    #expect(player.isSuspendedForTesting == true)
+    #expect(player.interruptedStationIdForTesting == "station-1")
+    #expect(player.wasPlayingBeforeInterruptionForTesting == true)
+  }
+
+  @Test("Pause during loading arms resume and clears the spinner state")
+  func pauseDuringLoadingArmsResumeAndClearsSpinner() {
+    let player = PlayolaStationPlayer(
+      fileDownloadManager: MockFileDownloadManager(), urlSession: MockURLSession())
+    player.configure(authProvider: MockAuthProvider())
+    player.setStateForTesting(.loading(0.5), stationId: "station-1")
+
+    player.pauseForInterruption()
+
+    if case .idle = player.state {
+    } else {
+      Issue.record("pause during loading must clear the spinner, got \(player.state)")
+    }
+    #expect(player.wasPlayingBeforeInterruptionForTesting == true)
+    #expect(player.interruptedStationIdForTesting == "station-1")
+  }
+}
